When I run 'make', the files generated in bin all have a second line that was the original first line when they were in src. Example: $ head -4 src/cegma.pl #!/usr/bin/perl
##########################################################################
#                                                                        #
$ head -4 bin/cegma
#!/usr/bin/perl -w
#!/usr/bin/perl

##########################################################################
$

While this does not appear to have any effect (at least on my machine), it is clearly not the intent to have this second line here.

This is due to the Makefile which is prepending $(PERL) to the files, instead of replacing the first line.
I fixed this by replacing:
@( echo "#!$(PERL) -w"; cat $< ) > $@;
with
sed '1c\#!$(PERL)' $< > $@

Note that the “-w” is not needed as all the Perl scripts “use warnings”.